Backend/ChatbotBackend/requirements.txt
Django==2.1.4
djangorestframework==3.9.0
django-cors-headers==2.5.2
numpy==1.15.4
pandas==0.23.4
scikit-learn==0.20.1
scipy==1.1.0
sklearn==0.0
spacy==2.1.3
pyspellchecker==0.4.0
pattern==3.6
